v1.0.8
Added a callback in order to let the user know when the video is (event->data ==== 1)started o (event->data === 0)finished.

Parameter passed now works, before the parameters were never assigned, so defaults was always setted.

Webpack Encore correctly support, now jQuery is detected correctly, and the function constructor is exported, in order to be imported in a proper way, as a module.